docs: Add more docs for google signer
