+++ src/services/google/constants.rs
@@ -1,1 +1,2 @@
+// Env values used in google services.
 pub const GOOGLE_APPLICATION_CREDENTIALS: &str = "GOOGLE_APPLICATION_CREDENTIALS";
