+++ src/services/google/signer.rs
@@ -2,7 +2,6 @@ use std::env;
 use std::ops::{Add, Sub};
 use std::sync::Arc;
 
-use crate::hash::base64_decode;
 use anyhow::{anyhow, Result};
 use http::{header, StatusCode};
 use jsonwebtoken::{Algorithm, EncodingKey, Header};
@@ -11,22 +10,12 @@ use reqwest::{Body, Request};
 use tokio::sync::RwLock;
 
 use super::constants::GOOGLE_APPLICATION_CREDENTIALS;
-use super::credential::{Claims, Credential, Token};
+use super::credential::{Claims, Credential, CredentialLoader, Token};
+use crate::hash::base64_decode;
 use crate::request::SignableRequest;
 use crate::time::{self, DateTime, Duration};
 
-enum CredentialLoader {
-    Path(String),
-    Content(String),
-    None,
-}
-
-impl Default for CredentialLoader {
-    fn default() -> Self {
-        CredentialLoader::None
-    }
-}
-
+/// Builder for Signer.
 #[derive(Default)]
 pub struct Builder {
     scope: Option<String>,
@@ -34,21 +23,44 @@ pub struct Builder {
 }
 
 impl Builder {
+    /// Specify scope url for Signer.
+    ///
+    /// For example, valid scopes for google cloud services should be
+    ///
+    /// - read-only: `https://www.googleapis.com/auth/devstorage.read_only`
+    /// - read-write: `https://www.googleapis.com/auth/devstorage.read_write`
+    /// - full-control: `https://www.googleapis.com/auth/devstorage.full_control`
+    ///
+    /// Reference: [Cloud Storage authentication](https://cloud.google.com/storage/docs/authentication)
     pub fn scope(&mut self, scope: &str) -> &mut Self {
         self.scope = Some(scope.to_string());
         self
     }
 
+    /// Load credential from path.
+    ///
+    /// The credential should be generated by Google Cloud Platform.
+    ///
+    /// Read more in [Authenticating as a service account](https://cloud.google.com/docs/authentication/production)
     pub fn credential_from_path(&mut self, path: &str) -> &mut Self {
         self.credential = CredentialLoader::Path(path.to_string());
         self
     }
 
+    /// Load credential from base64 content.
+    ///
+    /// The credential should be generated by Google Cloud Platform.
+    ///
+    /// Read more in [Authenticating as a service account](https://cloud.google.com/docs/authentication/production)
     pub fn credential_from_content(&mut self, credential: &str) -> &mut Self {
         self.credential = CredentialLoader::Content(credential.to_string());
         self
     }
 
+    /// Use exising information to build a new signer.
+    ///
+    ///
+    /// The builder should not be used anymore.
     pub async fn build(&mut self) -> Result<Signer> {
         let scope = match &self.scope {
             Some(v) => v.clone(),
@@ -80,6 +92,11 @@ impl Builder {
     }
 }
 
+/// Singer that implement Google OAuth2 Authentication.
+///
+/// ## Reference
+///
+/// -  [Authenticating as a service account](https://cloud.google.com/docs/authentication/production)
 pub struct Signer {
     scope: String,
     credential: Credential,
@@ -146,7 +163,41 @@ impl Signer {
         Ok(token)
     }
 
-    /// TODO: we can also send API via signed JWT: [Addendum: Service account authorization without OAuth](https://developers.google.com/identity/protocols/oauth2/service-account#jwt-auth)
+    /// Signing request.
+    ///
+    /// # Example
+    ///
+    /// ```no_run
+    /// use reqsign::services::google::Signer;
+    /// use reqwest::{Client, Request, Url};
+    /// use anyhow::Result;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<()>{
+    ///     // Signer will load region and credentials from environment by default.
+    ///     let signer = Signer::builder()
+    ///         .scope("https://www.googleapis.com/auth/devstorage.read_only")
+    ///         .credential_from_path("/path/to/credential/file")
+    ///         .build()
+    ///         .await?;
+    ///
+    ///     // Construct request
+    ///     let url = Url::parse("https://storage.googleapis.com/storage/v1/b/test")?;
+    ///     let mut req = reqwest::Request::new(http::Method::GET, url);
+    ///
+    ///     // Signing request with Signer
+    ///     signer.sign(&mut req).await?;
+    ///
+    ///     // Sending already signed request.
+    ///     let resp = Client::new().execute(req).await?;
+    ///     println!("resp got status: {}", resp.status());
+    ///     Ok(())
+    /// }
+    /// ```
+    ///
+    /// # TODO
+    ///
+    /// we can also send API via signed JWT: [Addendum: Service account authorization without OAuth](https://developers.google.com/identity/protocols/oauth2/service-account#jwt-auth)
     pub async fn sign(&self, req: &mut impl SignableRequest) -> Result<()> {
         let token = self.token().await?;
         req.apply_header(
